BAC-3400: rename file
